Reorganize to pull Python build/wrapper
