--- appveyor.yml
@@ -1,32 +1,0 @@
-branches:
-  only:
-  - master
-
-version: '0.1.0.{build}'
-
-platform:
-  - x64
-
-configuration:
-  - Release
-
-install:
-  - cmd: call C:\\Miniconda36-x64\Scripts\activate.bat
-  - cmd: conda config --set always_yes yes --set changeps1 no
-  - cmd: conda update -q conda
-  - cmd: conda config --add channels usgs-astrogeology
-  - cmd: conda config --add channels conda-forge
-  - cmd: conda install -c conda-forge numpy cython pytest
-  - cmd: conda install -c usgs-astrogeology libcsm cycsm
-  - cmd: conda install conda-build anaconda-client pytest-cov
-
-build_script:
-  - cmd: conda build .
-  - cmd: conda install --use-local usgscam
-
-test_script:
-  - cmd: pytest --cov=cycsm
-
-on_success:
-  - cmd: conda config --set anaconda_upload yes
-  - cmd: conda build --token %CONDA_UPLOAD_TOKEN% .
