--- recipe/README.md
@@ -1,5 +1,0 @@
-# Recipe
-
-This is an anaconda compliant recipe for building and install the CSM-CameraModel library.  This library depends upon the Community Sensor Model (CSM) libs and the  USGS CyCSM wrappers. These are automatically installed alongside the CSM-CameraModel library.   
-
-To install run: `conda build recipe` from the root project directory.
