--- recipe/bld.bat
@@ -1,3 +1,0 @@
-python setup.py install
-
-if errorlevel 1 exit 1
