--- recipe/build.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-if [ `uname` == Darwin ]; then
-  export MACOSX_DEPLOYMENT_TARGET=10.9;
-fi
-
-$PYTHON setup.py install;
