--- recipe/meta.yaml
@@ -1,39 +1,0 @@
-{% set version = "0.1.1" %}
-
-package:
-  name: usgscam
-  version: {{ version }}
-
-source:
-    git_url: https://github.com/USGS-Astrogeology/CSM-CameraModel
-    git_rev: master
-build:
-  number: 0
-
-detect_binary_files_with_prefix: true
-
-requirements:
-  build: 
-    - {{ compiler('cxx') }}
-    - vc 14  # [win and py>=35]
-  host:
-    - python
-    - setuptools
-    - numpy >=1.10
-    - cython >=0.27
-    - libcsm
-    - cycsm
-  run:
-    - python
-    - numpy >=1.10
-    - libcsm
-    - cycsm
-    - vc 14  # [win and py>=35]
-
-test:
-  imports:
-    - usgscam
-
-about:
-  home: https://github.com/USGS-Astrogeology/CSM-CameraModel
-  license: Unlicense
