--- tests/CW1071364100B_IU_5.json
@@ -1,86 +1,0 @@
-{
-    "boresight": [
-        0.0,
-        0.0,
-        1.0
-    ],
-    "ccd_center": [
-        512.5,
-        512.5
-    ],
-    "ephemeris_time": 482312178.16084045,
-    "focal_length": 78.25197917515999,
-    "focal_length_epsilon": 0.15,
-    "ifov": 179.6,
-    "instrument_id": "MDIS-WAC",
-    "itrans_line": [
-        0.0,
-        -0.0006761060916520159,
-        71.4399371332522
-    ],
-    "itrans_sample": [
-        0.0,
-        71.42857143,
-        0.0
-    ],
-    "kappa": -3.072868243248164,
-    "min_elevation": -100,
-    "max_elevation": 100,
-    "model_name": "USGS_ASTRO_FRAME_SENSOR_MODEL",
-    "nlines": 1024,
-    "nsamples": 1024,
-    "odt_x": [
-        0.0,
-        1.0,
-        0.0,
-        -7.720894252056575e-05,
-        3.599871902138938e-06,
-        0.0,
-        5.509035727272325e-06,
-        0.0,
-        5.509035727272406e-06,
-        0.0
-    ],
-    "odt_y": [
-        0.0,
-        0.0,
-        1.000000000026148,
-        0.0,
-        -7.720894252092194e-05,
-        3.599871782473616e-06,
-        0.0,
-        5.509035621941527e-06,
-        0.0,
-        5.5090308738198125e-06
-    ],
-    "omega": 2.9059307323725254,
-    "original_half_lines": 512.0,
-    "original_half_samples": 512.0,
-    "phi": 0.3032857092051579,
-    "pixel_pitch": 0.014,
-    "semi_major_axis": 2439.4,
-    "semi_minor_axis": 2439.4,
-    "spacecraft_name": "Messenger",
-    "starting_detector_line": 1.0,
-    "starting_detector_sample": 9.0,
-    "target_name": "Mercury",
-    "transx": [
-        0.0,
-        0.01399999999972,
-        0.0
-    ],
-    "transy": [
-        0.0,
-        1.32495711261385e-07,
-        0.013997772676294
-    ],
-    "x_sensor_origin": 28892.88978623381,
-    "x_sensor_velocity": 0,
-    "x_sun_position": 0,
-    "y_sensor_origin": 633940.116378508,
-    "y_sensor_velocity": 0,
-    "y_sun_position": 0,
-    "z_sensor_origin": 2683552.1331144567,
-    "z_sensor_velocity": 0,
-    "z_sun_position": 0
-}
