--- tests/cassini_nac.json
@@ -1,63 +1,0 @@
-{
-  "boresight": [0.0, 0.0, 1.0],
-   "ccd_center": [512.5, 512.5],
-   "ephemeris_time": 503818547.2384149,
-   "focal_length": 2003.44,
-   "focal_length_epsilon": 0.03,
-   "ifov": 6.0,
-   "instrument_id": "CASSINI_ISS_NAC",
-   "itrans_line": [0.0, 0.0, 83.33333333333333],
-   "itrans_sample": [0.0, 83.33333333333333, 0.0],
-   "kappa": -1.0344655958281868,
-   "max_elevation": 100,
-   "min_elevation": -100,
-   "model_name": "ISIS_MDISNAC_USGSAstro_1_Linux64_csm30.so",
-   "nlines": 1024,
-   "nsamples": 1024,
-   "omega": 0.7582919095772543,
-   "original_half_lines": 512.0,
-   "original_half_samples": 512.0,
-   "phi": 0.7767982642455958,
-   "pixel_pitch": 0.012,
-   "semi_major_axis": 256.6,
-   "semi_minor_axis": 251.4,
-   "spacecraft_name": "CASSINI",
-   "target_name": "ENCELADUS",
-   "transx": [0.0, 0.012, 0.0],
-   "transy": [0.0, 0.0, 0.012],
-   "x_sensor_origin": 8467607.57716877,
-   "x_sensor_velocity": -9453.44521955574,
-   "x_sun_position": 1039775477180.3715,
-   "y_sensor_origin": 940810.1674152733,
-   "y_sensor_velocity": -1496.3339337677503,
-   "y_sun_position": 860644074787.2188,
-   "z_sensor_origin": -5007748.542046298,
-   "z_sensor_velocity": -278.8874775806509,
-   "z_sun_position": 648585685779.2721,
-	"odt_x": [
-		0.0,
-		1.0,
-		0.0,
-		0.0,
-		0.0,
-		0.0,
-		0.0,
-		0.0,
-		0.0,
-		0.0
-	],
-	"odt_y": [
-		0.0,
-		0.0,
-		1.0,
-		0.0,
-		0.0,
-		0.0,
-		0.0,
-		0.0,
-		0.0,
-		0.0
-	],
-	"starting_detector_line": 0.0,
-	"starting_detector_sample": 0.0
-}
