--- tests/nac_state.json
@@ -1,132 +1,0 @@
-{
-    "m_boresight": [
-        0.0,
-        0.0,
-        1
-    ],
-    "m_ccdCenter": [
-        512.5,
-        512.5
-    ],
-    "m_currentParameterCovariance": [
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0
-    ],
-    "m_currentParameterValue": [
-        1728181.03407928,
-        -2088202.59129735,
-        2082707.60899282,
-        2.25613869898165,
-        0.0943461617903765,
-        -0.963008015000929
-    ],
-    "m_ephemerisTime": 418855170.49265,
-    "m_focalLength": 549.23479652106,
-    "m_focalLengthEpsilon": 0.5,
-    "m_iTransL": [
-        0.0,
-        0.0,
-        71.42857143
-    ],
-    "m_iTransS": [
-        0.0,
-        71.42857143,
-        0.0
-    ],
-    "m_ifov": 25.44,
-    "m_instrumentID": "MDIS-NAC",
-    "m_line_pp": 0.0,
-    "m_majorAxis": 2439400,
-    "m_minorAxis": 2439400,
-    "m_minElevation": -100,
-    "m_maxElevation": 100,
-    "m_nLines": 1024,
-    "m_nSamples": 1024,
-    "m_odtX": [
-        0.0,
-        1.0018542696238,
-        0.0,
-        0.0,
-        -0.000509444047494111,
-        0.0,
-        1.00401047146886e-05,
-        0.0,
-        1.00401047146886e-05,
-        0.0
-    ],
-    "m_odtY": [
-        0.0,
-        0.0,
-        1,
-        0.000906001059499675,
-        0.0,
-        0.000357484262662076,
-        0.0,
-        1.00401047146886e-05,
-        0.0,
-        1.00401047146886e-05
-    ],
-    "m_originalHalfLines": 512,
-    "m_originalHalfSamples": 512,
-    "m_pixelPitch": 0.014,
-    "m_sample_pp": 0.0,
-    "m_spacecraftName": "Messenger",
-    "m_spacecraftVelocity": [
-        0.0,
-        0.0,
-        0.0
-    ],
-    "m_startingDetectorLine": 1,
-    "m_startingDetectorSample": 9,
-    "m_sunPosition": [
-        -31648725087.5887,
-        -60633907522.7286,
-        -38729485.7733473
-    ],
-    "m_targetName": "Mercury",
-    "m_transX": [
-        0.0,
-        0.014,
-        0.0
-    ],
-    "m_transY": [
-        0.0,
-        0.0,
-        0.014
-    ],
-    "model_name": "USGS_ASTRO_FRAME_SENSOR_MODEL"
-}
